capstone-draft/etl/requirements.txt
redis
pyspark==3.4.0
requests
pandas
tqdm
python-decouple
